+++ app/routes/_index.tsx
@@ -6,8 +6,8 @@ import VeilederHilsen from '../komponenter/veilederhilsen/veilederhilsen';
 import { ESanitySteg } from '~/typer/sanity/sanity';
 import TekstBlokk from '~/komponenter/tekstBlokk/tekstBlokk';
 import { TypografiTyper } from '~/typer/typografi';
-import { useTekster } from '~/utils/sanityLoader';
 import SamtykkePanel from '~/komponenter/SamtykkePanel';
+import { useTekster } from '~/hooks/contextHooks';
 
 export const meta: V2_MetaFunction = () => {
   return [
