+++ app/utils/sanityLoader.ts
@@ -1,7 +1,5 @@
-import { useOutletContext } from '@remix-run/react';
 import { createClient } from '@sanity/client';
 import {
-  AppContext,
   ESanitySteg,
   ITekstinnhold,
   SanityDokument,
@@ -35,8 +33,3 @@ const strukturerInnholdForSteg = (
     .reduce((acc, dok) => {
       return { ...acc, [dok.api_navn]: dok };
     }, {});
-
-export function useTekster(steg: ESanitySteg) {
-  const { sanityTekster } = useOutletContext<AppContext>();
-  return sanityTekster[steg];
-}
